+++ lib/src/outbox/outbox-listener.ts
@@ -6,6 +6,7 @@ import {
   createLogicalReplicationListener,
 } from '../replication/logical-replication-listener';
 import { defaultConcurrencyStrategy } from '../strategies/concurrency-strategy';
+import { defaultListenerRestartTimeStrategy } from '../strategies/listener-restart-time-strategy';
 import { defaultMessageProcessingTimeoutStrategy } from '../strategies/message-processing-timeout-strategy';
 
 export type OutboxConfig = TransactionalOutboxInboxConfig;
@@ -50,4 +51,7 @@ const applyDefaultStrategies = (
   messageProcessingTimeoutStrategy:
     strategies?.messageProcessingTimeoutStrategy ??
     defaultMessageProcessingTimeoutStrategy(config),
+  listenerRestartTimeStrategy:
+    strategies?.listenerRestartTimeStrategy ??
+    defaultListenerRestartTimeStrategy(config),
 });